# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EstoqueClinic API",
    "version": "1.0.0",
    "description": "API REST para gestão de estoque de clínicas estéticas com multi-tenancy",
    "contact": {
      "name": "EstoqueClinic Support",
      "email": "suporte@estoqueclinic.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento Local"
    },
    {
      "url": "https://api.estoqueclinic.com",
      "description": "Produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT de autenticação. Header: Authorization: Bearer {token}"
      }
    },
    "schemas": {
      "ClinicGroup": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "name": {"type": "string"},
          "code": {"type": "string"},
          "description": {"type": "string"},
          "admin_email": {"type": "string", "format": "email"},
          "admin_phone": {"type": "string"},
          "subscription_plan": {"type": "string", "enum": ["basic", "premium", "enterprise"]},
          "subscription_expires_at": {"type": "string", "format": "date-time"},
          "max_clinics": {"type": "integer"},
          "max_users": {"type": "integer"},
          "settings": {"type": "object"},
          "is_active": {"type": "boolean"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "Clinic": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "clinic_group_id": {"type": "integer"},
          "name": {"type": "string"},
          "cnpj": {"type": "string"},
          "cro_number": {"type": "string"},
          "address_street": {"type": "string"},
          "address_number": {"type": "string"},
          "address_complement": {"type": "string"},
          "address_neighborhood": {"type": "string"},
          "address_city": {"type": "string"},
          "address_state": {"type": "string"},
          "address_zipcode": {"type": "string"},
          "phone": {"type": "string"},
          "email": {"type": "string", "format": "email"},
          "website": {"type": "string"},
          "logo_url": {"type": "string"},
          "timezone": {"type": "string"},
          "settings": {"type": "object"},
          "is_active": {"type": "boolean"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {"type": "string"},
          "message": {"type": "string"},
          "code": {"type": "integer"}
        }
      }
    }
  },
...
  "paths": {
...
    "/api/clinic-groups": {
      "get": {
        "tags": ["Clinic Groups"],
        "summary": "Listar grupos de clínicas",
        "description": "Lista todos os grupos de clínicas (apenas super admin)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 10}},
          {"name": "search", "in": "query", "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "Lista de grupos de clínicas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "clinic_groups": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/ClinicGroup"}
                    },
                    "pagination": {"type": "object"}
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"}
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Clinic Groups"],
        "summary": "Criar grupo de clínicas",
        "description": "Cria um novo grupo de clínicas (apenas super admin)",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "code", "admin_email"],
                "properties": {
                  "name": {"type": "string"},
                  "code": {"type": "string"},
                  "description": {"type": "string"},
                  "admin_email": {"type": "string", "format": "email"},
                  "admin_phone": {"type": "string"},
                  "subscription_plan": {"type": "string"},
                  "subscription_expires_at": {"type": "string", "format": "date-time"},
                  "max_clinics": {"type": "integer"},
                  "max_users": {"type": "integer"},
                  "settings": {"type": "object"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Grupo criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"},
                    "clinic_group": {"$ref": "#/components/schemas/ClinicGroup"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/clinics": {
      "get": {
        "tags": ["Clinics"],
        "summary": "Listar clínicas do grupo",
        "description": "Lista todas as clínicas do grupo do usuário",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 10}},
          {"name": "search", "in": "query", "schema": {"type": "string"}},
          {"name": "is_active", "in": "query", "schema": {"type": "boolean"}}
        ],
        "responses": {
          "200": {
            "description": "Lista de clínicas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "clinics": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/Clinic"}
                    },
                    "pagination": {"type": "object"}
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Clinics"],
        "summary": "Criar clínica",
        "description": "Cria uma nova clínica no grupo do usuário",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {"type": "string"},
                  "cnpj": {"type": "string"},
                  "cro_number": {"type": "string"},
                  "address_street": {"type": "string"},
                  "address_number": {"type": "string"},
                  "address_complement": {"type": "string"},
                  "address_neighborhood": {"type": "string"},
                  "address_city": {"type": "string"},
                  "address_state": {"type": "string"},
                  "address_zipcode": {"type": "string"},
                  "phone": {"type": "string"},
                  "email": {"type": "string", "format": "email"},
                  "website": {"type": "string"},
                  "logo_url": {"type": "string"},
                  "timezone": {"type": "string"},
                  "settings": {"type": "object"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Clínica criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"},
                    "clinic": {"$ref": "#/components/schemas/Clinic"}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}